i hate html

+++ src/main/java/web/dao/UserDaoImpl.java
@@ -1,7 +1,6 @@
 package web.dao;
 
 import org.springframework.stereotype.Repository;
-import org.springframework.transaction.annotation.Transactional;
 import web.model.User;
 
 import javax.persistence.EntityManager;
@@ -10,42 +9,50 @@ import java.util.List;
 
 @Repository
 public class UserDaoImpl implements UserDao {
+
     @PersistenceContext
     private EntityManager entityManager;
 
-    @Transactional
+
     @Override
-    public void addUser(User user) {
+    public void saveUser(User user) {
         entityManager.persist(user);
     }
 
-    @Transactional
     @Override
-    public List<User> getUsers() {
-        return entityManager.createQuery("SELECT u FROM User u", User.class).getResultList();
-    }
-
-    @Transactional(readOnly = true)
-    @Override
-    public User getUserById(int id) {
+    public User getUserById(Long id) {
         return entityManager.find(User.class, id);
     }
 
-    @Transactional
     @Override
     public void updateUser(User user) {
-        User userToBeUpdate = entityManager.find(User.class, user.getId());
-        userToBeUpdate.setName(user.getName());
-        userToBeUpdate.setLastName(user.getLastName());
-        userToBeUpdate.setAge(user.getAge());
+        entityManager.merge(user);
     }
 
-    @Transactional
     @Override
-    public void deleteUser(int id) {
-        User user = entityManager.find(User.class, id);
-        entityManager.remove(user);
+    public void deleteUser(Long id) {
+        try {
+            User user = entityManager.find(User.class, id);
+            if (user != null) {
+                entityManager.remove(user);
+            }
+        } catch (NullPointerException e) {
+            System.out.println("User с указанным вами id не существует!");
+        }
+    }
 
+    @Override
+    public List<User> getAllUsers() {
+        return entityManager.createQuery("select u from User u", User.class)
+                .getResultList();
+    }
 
+    @Override
+    public User getUserByName(String username) {
+        return entityManager.createQuery("select u from User u where u.username = :username", User.class)
+                .setParameter("username", username)
+                .getSingleResult();
     }
 }
+
+
